ver 1.2.9.2
Add Vampire Skill

+++ Assets/Scripts/CharactersHealth.cs
@@ -37,7 +37,7 @@ public class CharactersHealth : MonoBehaviour
         }
     }
 
-    public void TakePills(int pills)
+    public void TakePills(float pills)
     {
         if(pills >= 0)
         {
